add: test for /user & /new-user & /userlist

+++ server/controllers/user.controller.ts
@@ -6,14 +6,13 @@ import { Document, Error } from "mongoose";
 export const getUser = async (ctx:Context) => {
   try {
     const id = ctx.request.body?.id;
-    if(id) {
-      console.log('Getting user: ', id);
-      let user = await User.findOne({ _id: id });
-      ctx.body = user;
-      ctx.status = 200;
-    } else {
-      throw new Error('User id is not provided.')
-    }
+    if(!id) throw new Error('User id is not provided')
+
+    console.log('Getting user: ', id);
+    let user = await User.findOne({ _id: id });
+    ctx.body = user;
+    ctx.status = 200;
+
   } catch (err) {
     console.log("ERROR: ", err);
     ctx.status = 400;
@@ -23,15 +22,16 @@ export const getUser = async (ctx:Context) => {
 export const postUser = (ctx:Context) => {
   try {
     const user = ctx.request.body;
-    if(user) {
-      console.log("Creating new user:", user);
-      const newUser = new User(user);
-      newUser.save();
-      console.log("Inserted _id is ", newUser._id.toString());
-      ctx.status = 201;
-    } else {
-      throw new Error('User data is not provided')
-    }
+
+    if(!user || Object.keys(user).length === 0 ) throw new Error('User data is not provided')
+
+    const newUser = new User(user);
+
+    newUser.save();
+    console.log("Inserted _id is ", newUser._id.toString());
+    ctx.body = newUser;
+    ctx.status = 201;
+
   } catch (err) {
     console.log("ERROR: ", err);
     ctx.status = 400;
@@ -63,17 +63,16 @@ export const getAllUsers = async (ctx:Context) => {
 export const postUserList = async (ctx:Context) => {
   try {
     const newUser = ctx.request.body;
-    if(newUser) {
-      const user = await User.findOne({ _id:newUser._id });
-      // add this condition to prevent user === undefined
-      if(user) {
-        user['contacts'] = newUser.contacts as string[];
-        user.save();
-      }
-      ctx.status = 201;
-    } else {
-      throw new Error('User id is not provided')
+    if(!newUser || Object.keys(newUser).length === 0) throw new Error('User id is not provided')
+
+    const user = await User.findOne({ _id:newUser._id });
+    // add this condition to prevent user === undefined
+    if(user) {
+      user['contacts'] = newUser.contacts as string[];
+      user.save();
     }
+    ctx.status = 201;
+
   } catch (err) {
     console.log('Error in the server postUserList: ', err);
     ctx.status = 400;
@@ -84,6 +83,7 @@ export const deleteContacts = async (ctx:Context) => {
   try {
     const userId:string = ctx.request.body?.id as string;
     // what is ObjectId? 
+    if(!userId) throw new Error('No user id provided')
     const user = await User.findOne({ '_id': new ObjectId(userId) });
     // add this condition to prevent user === undefined
     if(user) {
@@ -101,6 +101,7 @@ export const deleteContacts = async (ctx:Context) => {
 export const deleteUser = (ctx:Context) => {
   try {
     const userId:string = ctx.request.body?.id as string;
+    if(!userId) throw new Error('No user id provided')
     console.log('User ID to delete is: ', userId);
     User.findOneAndDelete({ "_id": new ObjectId(userId) }, function(err:Error, docs:Document) {
       if (err) {
